Moved MetricsService to its own file
